feat: enhance order and inventory handling with minor unit adjustments and new configurations
- Updated OrderConfig to include ShowVirtualStockRemark for user-side display of virtual product remarks.
- Added migration for converting legacy decimal money values to minor-unit int64 values across products, orders, and promo codes.
- Adjusted analytics and dashboard handlers to use int64 for monetary values, ensuring consistency in data types.
- Enhanced product and promo code handlers to accept minor unit values for pricing and discounts.
- Introduced source filtering in inventory log queries for better log management.
- Removed deprecated API order handler to streamline the codebase.

+++ backend/internal/service/promo_code_service.go
@@ -83,7 +83,7 @@ func (s *PromoCodeService) Delete(id uint) error {
 }
 
 // ValidateCode 验证优惠码是否可用于指定商品
-func (s *PromoCodeService) ValidateCode(code string, productIDs []uint, orderAmount float64) (*models.PromoCode, float64, error) {
+func (s *PromoCodeService) ValidateCode(code string, productIDs []uint, orderAmount int64) (*models.PromoCode, int64, error) {
 	code = strings.ToUpper(strings.TrimSpace(code))
 	promoCode, err := s.repo.FindByCode(code)
 	if err != nil {
